net/mana: start secondary process queues by default
Secondary processes are started after primary, and in most cases with
the device already started. Make them being able to process packets as
soon as they start.

This also works with the case where the primary process decides to start
the device at a later time after secondary processes have started. The
application should guarantee not to send any packets before the device is
started.

Signed-off-by: Long Li <longli@microsoft.com>