net/bnxt: update consumer index of NQ regularly
Update the consumer index of the default CQ/NQ at regular intervals.
Since the size of the queue can vary, ring the queue at 1/8th the queue
depth.

Also skip NQ doorbell update if done already.

If the NQ is rearmed already or if there was no valid entry,
which is indicated by cnt being 0, skip ringing the doorbell for the NQ.

Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>
Reviewed-by: Damodharam Ammepalli <damodharam.ammepalli@broadcom.com>
Reviewed-by: Somnath Kotur <somnath.kotur@broadcom.com>
Reviewed-by: Kalesh AP <kalesh-anakkur.purayil@broadcom.com>